Histo-TransCLIP-main/requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blobfile==3.0.0
braceexpand==0.1.7
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
cmake==4.0.2
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
docker-pycreds==0.4.0
ecos==2.0.14
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fairscale==0.4.13
filelock==3.18.0
flake8==7.2.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
ftfy==6.3.1
future==1.0.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
h5py==3.13.0
huggingface-hub==0.31.2
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
isort==4.3.21
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.14.3
lit==18.1.8
littleutils==0.2.4
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
munkres==1.1.4
mypy==1.15.0
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.6.77
ogb==1.3.6
open_clip_torch==2.32.0
opencv-python==4.11.0.86
openpyxl==3.1.5
osqp==0.6.7.post3
outdated==0.2.2
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
pyarrow==20.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pycodestyle==2.13.0
pycryptodomex==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
pyDeprecate==0.3.2
pyflakes==3.3.2
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.1
pytz==2025.2
PyYAML==6.0.2
qdldl==0.1.7.post5
regex==2024.11.6
requests==2.32.3
rouge==1.0.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scikit-survival==0.24.1
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.29.1
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
soupsieve==2.7
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==0.9.8
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.0
torchaudio==2.7.0
torchmetrics==0.7.3
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.11
wcwidth==0.2.13
webdataset==0.2.111
wilds==1.2.2
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.20.0
